+++ src/Collection/test.js
@@ -4,6 +4,8 @@ import { noop } from '../utils/fp'
 import Query from '../Query'
 import * as Q from '../QueryDescription'
 import { logger } from '../utils/common'
+import { toPromise } from '../utils/fp/Result'
+
 import { mockDatabase, MockTask, testSchema } from '../__tests__/testModels'
 
 import { CollectionChangeTypes } from './common'
@@ -42,7 +44,9 @@ describe('finding records', () => {
 
     // TODO: Don't mock
     // TODO: Should ID (not raw) response be tested?
-    adapter.find = jest.fn().mockReturnValueOnce({ id: 'm1' })
+    adapter.find = jest
+      .fn()
+      .mockImplementation((table, id, callback) => callback({ value: { id: 'm1' } }))
 
     // calls db
     const m1 = await collection.find('m1')
@@ -54,7 +58,7 @@ describe('finding records', () => {
     expect(collection._cache.map.size).toBe(1)
 
     // check call
-    expect(adapter.find.mock.calls[0]).toEqual(['mock_tasks', 'm1'])
+    expect(adapter.find.mock.calls[0]).toEqual(['mock_tasks', 'm1', expect.anything()])
 
     // second find will be from cache
     const m1Cached = await collection.find('m1')
@@ -65,13 +69,12 @@ describe('finding records', () => {
   })
   it('rejects promise if record cannot be found', async () => {
     const { tasks: collection, adapter } = mockDatabase()
-
-    adapter.find = jest.fn().mockReturnValue(null)
+    const findSpy = jest.spyOn(adapter, 'find')
 
     await expect(collection.find('m1')).rejects.toBeInstanceOf(Error)
     await expect(collection.find('m1')).rejects.toBeInstanceOf(Error)
 
-    expect(adapter.find.mock.calls.length).toBe(2)
+    expect(findSpy.mock.calls.length).toBe(2)
   })
 })
 
@@ -79,12 +82,14 @@ describe('fetching queries', () => {
   it('fetches queries and caches records', async () => {
     const { tasks: collection, adapter } = mockDatabase()
 
-    adapter.query = jest.fn().mockReturnValueOnce([{ id: 'm1' }, { id: 'm2' }])
+    adapter.query = jest
+      .fn()
+      .mockImplementation((query, cb) => cb({ value: [{ id: 'm1' }, { id: 'm2' }] }))
 
     const query = mockQuery(collection)
 
     // fetch, check models
-    const models = await collection.fetchQuery(query)
+    const models = await toPromise(callback => collection._fetchQuery(query, callback))
     expect(models.length).toBe(2)
 
     expect(models[0]._raw).toEqual({ id: 'm1' })
@@ -104,13 +109,13 @@ describe('fetching queries', () => {
   it('fetches query records from cache if possible', async () => {
     const { tasks: collection, adapter } = mockDatabase()
 
-    adapter.query = jest.fn().mockReturnValueOnce(['m1', { id: 'm2' }])
+    adapter.query = jest.fn().mockImplementation((query, cb) => cb({ value: ['m1', { id: 'm2' }] }))
 
     const m1 = new MockTask(collection, { id: 'm1' })
     collection._cache.add(m1)
 
     // fetch, check models
-    const models = await collection.fetchQuery(mockQuery(collection))
+    const models = await toPromise(cb => collection._fetchQuery(mockQuery(collection), cb))
     expect(models.length).toBe(2)
     expect(models[0]).toBe(m1)
     expect(models[1]._raw).toEqual({ id: 'm2' })
@@ -121,14 +126,16 @@ describe('fetching queries', () => {
   it('fetches query records from cache even if full raw object was sent', async () => {
     const { tasks: collection, adapter } = mockDatabase()
 
-    adapter.query = jest.fn().mockReturnValueOnce([{ id: 'm1' }, { id: 'm2' }])
+    adapter.query = jest
+      .fn()
+      .mockImplementation((query, cb) => cb({ value: [{ id: 'm1' }, { id: 'm2' }] }))
 
     const m1 = new MockTask(collection, { id: 'm1' })
     collection._cache.add(m1)
 
     // fetch, check if error occured
     const spy = jest.spyOn(logger, 'error').mockImplementation(() => {})
-    const models = await collection.fetchQuery(mockQuery(collection))
+    const models = await toPromise(cb => collection._fetchQuery(mockQuery(collection), cb))
     expect(spy).toHaveBeenCalledTimes(1)
     spy.mockRestore()
 
@@ -142,13 +149,13 @@ describe('fetching queries', () => {
 
     adapter.count = jest
       .fn()
-      .mockReturnValueOnce(5)
-      .mockReturnValueOnce(10)
+      .mockImplementationOnce((query, callback) => callback({ value: 5 }))
+      .mockImplementationOnce((query, callback) => callback({ value: 10 }))
 
     const query = mockQuery(collection)
 
-    expect(await collection.fetchCount(query)).toBe(5)
-    expect(await collection.fetchCount(query)).toBe(10)
+    expect(await toPromise(callback => collection._fetchCount(query, callback))).toBe(5)
+    expect(await toPromise(callback => collection._fetchCount(query, callback))).toBe(10)
 
     expect(adapter.count.mock.calls.length).toBe(2)
     expect(adapter.count.mock.calls[0][0]).toEqual(query.serialize())
@@ -173,7 +180,7 @@ describe('creating new records', () => {
     expect(m1._isCommitted).toBe(true)
     expect(newModelSpy).toHaveBeenCalledTimes(1)
     expect(dbBatchSpy).toHaveBeenCalledTimes(1)
-    expect(dbBatchSpy).toHaveBeenCalledWith([['create', 'mock_tasks', m1._raw]])
+    expect(dbBatchSpy).toHaveBeenCalledWith([['create', 'mock_tasks', m1._raw]], expect.anything())
     expect(observer).toHaveBeenCalledTimes(1)
     expect(observer).toHaveBeenCalledWith([{ record: m1, type: CollectionChangeTypes.created }])
     expect(collection._cache.get(m1.id)).toBe(m1)