+++ src/adapters/lokijs/WorkerBridge.js
@@ -1,22 +1,18 @@
 // @flow
 
+import type { ResultCallback } from '../../utils/fp/Result'
 import {
-  responseActions,
   type WorkerExecutorType,
   type WorkerResponse,
   type WorkerExecutorPayload,
-  type WorkerResponsePayload,
+  type WorkerResponseData,
 } from './common'
 
-type PromiseResponse = WorkerResponsePayload => void
 type WorkerAction = {
   id: number,
-  resolve: PromiseResponse,
-  reject: PromiseResponse,
+  callback: ResultCallback<WorkerResponseData>,
 }
 type WorkerActions = WorkerAction[]
-
-const { RESPONSE_SUCCESS, RESPONSE_ERROR } = responseActions
 
 function createWorker(useWebWorker: boolean): Worker {
   if (useWebWorker) {
@@ -43,33 +39,32 @@ class WorkerBridge {
   constructor(useWebWorker: boolean): void {
     this._worker = createWorker(useWebWorker)
     this._worker.onmessage = ({ data }) => {
-      const { type, payload, id: responseId }: WorkerResponse = (data: any)
-      const { resolve, reject, id } = this._pendingActions.shift()
+      const { result, id: responseId }: WorkerResponse = (data: any)
+      const { callback, id } = this._pendingActions.shift()
 
       // sanity check
       if (id !== responseId) {
-        reject((new Error('Loki worker responses are out of order'): any))
+        callback({ error: (new Error('Loki worker responses are out of order'): any) })
+        return
       }
 
-      if (type === RESPONSE_ERROR) {
-        reject(payload)
-      } else if (type === RESPONSE_SUCCESS) {
-        resolve(payload)
-      }
+      callback(result)
     }
   }
 
   // TODO: `any` return should be `WorkerResponsePayload`
-  send(
+  send<T>(
     type: WorkerExecutorType,
     payload: WorkerExecutorPayload = [],
+    callback: ResultCallback<T>,
     cloneMethod: 'shallowCloneDeepObjects' | 'immutable' | 'deep' = 'deep',
-  ): Promise<any> {
-    return new Promise((resolve, reject) => {
-      const id = nextActionId()
-      this._pendingActions.push({ resolve, reject, id })
-      this._worker.postMessage({ id, type, payload, cloneMethod })
+  ): void {
+    const id = nextActionId()
+    this._pendingActions.push({
+      callback: (callback: any),
+      id,
     })
+    this._worker.postMessage({ id, type, payload, cloneMethod })
   }
 }